requirements.txt
git+http://github.com/django-nonrel/djangotoolbox@toolbox-1.4

git+http://github.com/django-nonrel/djangoappengine@appengine-1.4

git+http://github.com/django-nonrel/django-dbindexer@dbindexer-1.4

git+http://github.com/django-nonrel/django@nonrel-1.4

hg+http://bitbucket.org/twanschik/django-autoload
git+http://github.com/chrisns/SpiffWorkflow
git+http://github.com/chrisns/gvgen
GoogleAppEngineCloudStorageClient==1.9.0.0

Markdown==2.4.1
NoseGAE==0.2.1
PyYAML==3.11
Pygments==1.6
WebOb==1.4
WebTest==2.0.15
argparse==1.2.1
astroid==1.1.1
beautifulsoup4==4.3.2
coverage==3.7.1
defusedxml==0.4.1
django-appengine-toolkit==0.2.1
django-filter==0.7
django-guardian==1.2.0
djangorestframework==2.3.13
git+https://github.com/ze-phyr-us/linkenv.git
logilab-common==0.61.0
mock==1.0.1
nose==1.3.3
pylint==1.2.1
six==1.6.1
waitress==0.8.9
wsgiref==0.1.2
jsonschema==2.3.0
django-picklefield=0.3.1